backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery==5.4.0
redis==5.0.8
slowapi==0.1.9
imagekitio==3.2.0
python-multipart==0.0.10
email-validator==2.2.0
httpx==0.27.2

# Phase 2+ deps baked in now to avoid image rebuilds
pypdf[crypto]==5.0.1
pdf2image==1.17.0
reportlab==4.2.5
opencv-python-headless==4.10.0.84
Pillow==10.4.0
pytesseract==0.3.13
